feat: 예약이 존재하는 시간을 삭제할 때 예외 처리

+++ src/main/java/roomescape/service/TimeSlotService.java
@@ -3,17 +3,22 @@ package roomescape.service;
 import java.time.LocalTime;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
+import roomescape.model.Reservation;
 import roomescape.model.TimeSlot;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.TimeSlotRepository;
 
 @Service
 public class TimeSlotService {
 
+    private final ReservationRepository reservationRepository;
     private final TimeSlotRepository timeSlotRepository;
 
     @Autowired
-    public TimeSlotService(TimeSlotRepository timeSlotRepository) {
+    public TimeSlotService(ReservationRepository reservationRepository, TimeSlotRepository timeSlotRepository) {
+        this.reservationRepository = reservationRepository;
         this.timeSlotRepository = timeSlotRepository;
     }
 
@@ -28,6 +33,10 @@ public class TimeSlotService {
     }
 
     public boolean removeById(long id) {
+        List<Reservation> reservations = reservationRepository.findByTimeSlotId(id);
+        if (!reservations.isEmpty()) {
+            throw new IllegalStateException("삭제하려는 예약 시간을 사용하는 예약이 있습니다.");
+        }
         return timeSlotRepository.removeById(id);
     }
 }
